Move the LinkedClass printer in a separate utility object.
